+++ src/main/java/com/example/demo/domain/model/Employee.java
@@ -1,21 +1,14 @@
 package com.example.demo.domain.model;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Transient;
-
 import lombok.Data;
 
 @Data
-@Entity
 public class Employee {
 
-	@Id
 	private Integer id;
 	private String name;
 	private Integer age;
 	private Integer gender;
-	@Transient
 	private String genderString;
 
 	/**
