+++ src/main/java/com/example/demo/listener/WriteListener.java
@@ -1,8 +1,8 @@
 package com.example.demo.listener;
 
-import java.util.List;
 
 import org.springframework.batch.core.ItemWriteListener;
+import org.springframework.batch.item.Chunk;
 import org.springframework.stereotype.Component;
 
 import com.example.demo.domain.model.Employee;
@@ -14,17 +14,17 @@ import lombok.extern.slf4j.Slf4j;
 public class WriteListener implements ItemWriteListener<Employee> {
 
 	@Override
-	public void beforeWrite(List<? extends Employee> items) {
-		// Do nothing
+	public void beforeWrite(Chunk<? extends Employee> items) {
+		// Do Nothing
 	}
 
 	@Override
-	public void afterWrite(List<? extends Employee> items) {
+	public void afterWrite(Chunk<? extends Employee> items) {
 		log.debug("AfterWrite: count={}", items.size());
 	}
 
 	@Override
-	public void onWriteError(Exception exception, List<? extends Employee> items) {
+	public void onWriteError(Exception exception, Chunk<? extends Employee> items) {
 		log.error("WriteError: errorMessage={}", exception.getMessage(), exception);
 	}
 }
