+++ src/main/resources/application.yml
@@ -1,7 +1,7 @@
 spring:
   batch:
     job:
-      names: ExportJdbcPagingJob
+      names: exportJdbcCursorJob
     jdbc:
       initialize-schema: always
       schema: classpath:org/springframework/batch/core/schema-postgresql.sql
